Further simplify auth code.
Make the base use-case an interface. This removes any thread handling code which will fix a couple problems. With the previous setup, chaining use-cases would create a mess. Because, every use-case has a subscribeOn and an observeOn operator. Also, we weren't able to change/choose the SchedulerProvider, and the only use for it is to facilitate testing. This is, of course, a code smell.