Show post's subreddit icon
When loading, display a placeholder as a subreddit icon. The placeholder is a solid color. Its color comes from the subreddit's primary/key color.
If the subreddit has no icon, use a default icon which has a background color set to the subreddit's primary/key color.
A default color is used when a subreddit has no primary/key color, this applies to both the placeholder and the default subreddit's icon background.